+++ reservation-api/src/main/java/ca/ulaval/glo4002/reservation/domain/fullcourse/recipe/vegan/VeganDessert.java
@@ -1,11 +1,13 @@
 package ca.ulaval.glo4002.reservation.domain.fullcourse.recipe.vegan;
 
+import java.math.BigDecimal;
+
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Ingredient;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.IngredientName;
 import ca.ulaval.glo4002.reservation.domain.fullcourse.Recipe;
 
 public class VeganDessert extends Recipe {
-  private static final double WORCESTERSHIRE_SAUCE_QUANTITY = 5;
+  private static final BigDecimal WORCESTERSHIRE_SAUCE_QUANTITY = BigDecimal.valueOf(5);
 
   public VeganDessert() {
     ingredients.add(new Ingredient(IngredientName.WORCESTERSHIRE_SAUCE,
