+++ reservation-api/src/test/java/ca/ulaval/glo4002/reservation/domain/reservation/ReservationIngredientCalculatorTest.java
@@ -3,6 +3,7 @@ package ca.ulaval.glo4002.reservation.domain.reservation;
 import static com.google.common.truth.Truth.assertThat;
 import static org.mockito.BDDMockito.given;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -37,22 +38,22 @@ class ReservationIngredientCalculatorTest {
     Customer customer = new CustomerBuilder().withRestriction(RestrictionType.VEGAN).build();
     Table table = new TableBuilder().withCustomer(customer).build();
     Reservation reservation = new ReservationBuilder().withTable(table).build();
-    Map<IngredientName, Double> veganMenu = givenVeganCourseIngredientsQuantity();
+    Map<IngredientName, BigDecimal> veganMenu = givenVeganCourseIngredientsQuantity();
     given(menuRepository.getIngredientsQuantity(RestrictionType.VEGAN)).willReturn(veganMenu);
 
     // when
-    Map<IngredientName, Double> ingredientQuantity = reservationIngredientCalculator.getReservationIngredientsQuantity(reservation);
+    Map<IngredientName, BigDecimal> ingredientQuantity = reservationIngredientCalculator.getReservationIngredientsQuantity(reservation);
 
     // then
     assertThat(ingredientQuantity).isEqualTo(veganMenu);
   }
 
-  private Map<IngredientName, Double> givenVeganCourseIngredientsQuantity() {
-    Map<IngredientName, Double> ingredientDoubleMap = new HashMap<>();
-    ingredientDoubleMap.put(IngredientName.TOMATO, 5.0);
-    ingredientDoubleMap.put(IngredientName.KIWI, 8.0);
-    ingredientDoubleMap.put(IngredientName.WORCESTERSHIRE_SAUCE, 5.0);
-    ingredientDoubleMap.put(IngredientName.KIMCHI, 10.0);
+  private Map<IngredientName, BigDecimal> givenVeganCourseIngredientsQuantity() {
+    Map<IngredientName, BigDecimal> ingredientDoubleMap = new HashMap<>();
+    ingredientDoubleMap.put(IngredientName.TOMATO, BigDecimal.valueOf(5.0));
+    ingredientDoubleMap.put(IngredientName.KIWI, BigDecimal.valueOf(8.0));
+    ingredientDoubleMap.put(IngredientName.WORCESTERSHIRE_SAUCE, BigDecimal.valueOf(5.0));
+    ingredientDoubleMap.put(IngredientName.KIMCHI, BigDecimal.valueOf(10.0));
     return ingredientDoubleMap;
   }
 }
